integrates/05.flask-oauth-integration/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-session==0.5.0
authlib==1.3.0
openfga-sdk==0.4.0
python-dotenv==1.0.0
pyjwt==2.8.0
requests==2.31.0
